Preliminary work to for clang detection
